backend2/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.3.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.17.98
botocore==1.20.98
cached-property==1.5.2
celery==5.2.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
# cytoolz==0.11.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.1.7
Django==3.2.12
django-cors-headers==3.7.0
django-easy-health-check==0.1
django-ebhealthcheck==2.0.1
django-filter==2.4.0
django-import-export==2.5.0
django-oauth-toolkit==1.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-flex-fields==0.9.3
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==2.10
importlib-metadata==4.8.1
ipfshttpclient==0.8.0a2
jmespath==0.10.0
jsonschema==3.2.0
jwcrypto==1.0
kombu==5.2.3
lru-dict==1.1.7
magic-admin==0.0.5
MarkupPy==1.14
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
parsimonious==0.8.1
Pillow==9.0.1
prompt-toolkit==3.0.20
protobuf==3.19.4
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.14.1
PyJWT==2.1.0
pyrsistent==0.18.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.2
stripe==2.58.0
tablib==3.0.0
toolz==0.11.2
typing-extensions==3.10.0.2
urllib3==1.26.5
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.6.0
